src/requirements.txt
crypten==0.4.0
sktime==0.5.3
numpy==1.19.5
tensorflow_gpu==2.4.1
nest_asyncio==1.5.1
scikit_learn==0.24.1
tensorflow_privacy==0.5.1
torch==1.7.1
torchvision==0.8.2
tqdm==4.30.0